Refactored the player rendering into the map view
